feat: implementar microservicio Blacklist con validacion de RUT y excepciones

+++ manitasUnidas/manitasUnidas/blackList/src/main/java/manitasUnidas/blackList/Service/BlackListService.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import manitasUnidas.blackList.Model.BlackList;
 import manitasUnidas.blackList.Repository.BlackListRepository;
+import manitasUnidas.blackList.exception.ResourceNotFoundException;
 @Service
 public class BlackListService {
     @Autowired
@@ -17,20 +18,40 @@ public class BlackListService {
     }
 
     public BlackList bloquearUsuario(BlackList registro) {
-        if (repository.existsByRut(registro.getRut())) {
-            throw new RuntimeException("Este RUT ya se encuentra en la lista negra.");
-        }
-        registro.setFechaBloqueo(LocalDate.now());
-        return repository.save(registro);
+    if (repository.existsByRut(registro.getRut())) {
+        // El Handler lo capturará y enviará un 400
+        throw new RuntimeException("El RUT " + registro.getRut() + " ya está en la lista negra.");
+    }
+    registro.setFechaBloqueo(LocalDate.now());
+    return repository.save(registro);
     }
 
     public BlackList buscarPorRut(String rut) {
-        return repository.findByRut(rut)
-                .orElseThrow(() -> new RuntimeException("El RUT no está en la lista negra."));
+    return repository.findByRut(rut)
+            .orElseThrow(() -> new ResourceNotFoundException("El RUT " + rut + " no está sancionado."));
+    }
+
+    public BlackList obtenerPorId(Long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("No existe un registro en la lista negra con el ID: " + id));
     }
 
     public void desbloquear(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("No se puede eliminar: No existe el ID: + " + id);
+        }
         repository.deleteById(id);
+    }
+
+    public BlackList actualizarBlackList(Long id, BlackList datosNuevos) {
+        BlackList blackListExistente = obtenerPorId(id);
+
+        blackListExistente.setRut(datosNuevos.getRut());
+        blackListExistente.setAutorizadoPor(datosNuevos.getAutorizadoPor());
+        blackListExistente.setMotivo(datosNuevos.getMotivo());
+        blackListExistente.setFechaBloqueo(datosNuevos.getFechaBloqueo());
+
+        return repository.save(blackListExistente);
     }
 
 }
